Plug-And-Play Ranged Weapons (WIP)
Attempting to create a Plug-And-Play class system for easy addition of ranged weapons for employees who aren't necessarily programmers.

Refined Weapon further. Pretty limited, even as a virtual class. Really should just contain inventory data.

Ranged Weapon and Melee Weapon classes are being reworked to be as plug-and-play as possible. Requiring only a copying of the existing ranged/melee weapon template and some in-editor changes to function.

Additionally, added in the model for the 1813 army pistol and some pipes to mess around with.