Massively simplified firearms through animation player and tree.
Mainly worked on RangedWeapon and added in 1795 Musket. Still needs a lot of refinement but use of the animation player and tree has allowed for greatly simplified code.

Having some issues with particles, in particular the smoke which was an issue with the test double barrel as well

glitch when it comes to handling types for Capybara AI. This will likely be the next thing I'm working on so not worreid.