Skinning Done, Menu and Settings Started
Skinning is on hold for the time being until we want to spend time refining, it works but the math and scoring need tuning and adjustment.

Main Menu has a base layout now at least with some menu and tab switching logic for settings.

Settings singleton has been created to load from settings.ini user file (or generate one if not present) and fiddle with that. Needs some smartening up and more settings implementations.